Solution for Day One
I went with a memory intensive approach, rather than a time intensive
approach. I iterate through the array, checking to see if past numbers
can be used to sum to the desired amount. This lets me iterate only once
through the array (besides converting it to numbers). I could definitely
clean it up and also convert to integers at the time of iteration as
well.